Fix modules and sortable library functionallity
