+++ src/index.js
@@ -1,9 +1,8 @@
 import './style.css';
-import Tasks from './modules/tasks.js';
-import Sortable from './modules/Sortable/Sortable.js';
+import TasksList from './modules/taskslist.js';
 
 const inTsk = {};
-const objTasks = new Tasks();
+const objTasks = new TasksList();
 
 if (localStorage.savedTasks) {
   objTasks.tasks = JSON.parse(localStorage.getItem('savedTasks'));
@@ -24,7 +23,8 @@ btnAddTasks.addEventListener('click', () => {
     inTsk.description = inputElement.value;
     inTsk.completed = false;
     inTsk.index = objTasks.tasks.length;
-    objTasks.addTask(new Tasks(inTsk.description, inTsk.completed, inTsk.index));
+    objTasks.addTask(new TasksList(inTsk.description, inTsk.completed, inTsk.index));
+    objTasks.populateFields();
     inputElement.value = '';
   }
 });
@@ -37,7 +37,8 @@ inputElement.addEventListener('keyup', (e) => {
       inTsk.description = inputElement.value;
       inTsk.completed = false;
       inTsk.index = objTasks.tasks.length;
-      objTasks.addTask(new Tasks(inTsk.description, inTsk.completed, inTsk.index));
+      objTasks.addTask(new TasksList(inTsk.description, inTsk.completed, inTsk.index));
+      objTasks.populateFields();
       inputElement.value = '';
     }
   }
@@ -49,11 +50,6 @@ btnClearTasks.addEventListener('click', () => {
   objTasks.populateFields();
   ulElement.innerHTML = '';
   root.append(objTasks.displayTasks(), btnClearTasks);
-});
-
-new Sortable(ulElement, {
-  animation: 350,
-  fallbackOnBody: false,
 });
 
 objTasks.displayTasks();
